MLDL/langgraph_supervisor_test/pyproject.toml
[project]
name = "langgraph-supervisor-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-openai>=0.3.32",
    "langgraph-supervisor>=0.0.29",
    "langsmith>=0.4.21",
    "python-dotenv>=1.1.1",
]
